Move auth and authz to the new module
