+++ common/errors/errors.go
@@ -1,1 +1,6 @@
+
 package errors
+
+import "errors"
+
+var ErrForbidden = errors.New("forbidden")
